refactor(home): split the three oversized component files along seams
Same-package splits, so no import changes anywhere:

- MusicCard.kt (624 -> 201 lines): meta block to MusicCardMeta.kt,
  progress/transport to MusicCardTransport.kt, connect/empty states to
  MusicCardStates.kt. The states gain a Music prefix
  (MusicConnectState/MusicEmptyState) because package-visible names must
  not collide with WeatherCard's private EmptyState.
- MapPanel.kt (620 -> 275 lines): MapConfig + MAP_ZOOM to MapConfig.kt;
  the whole snapshot backend (SnapshotMap, MapSnapshotter.render, camera
  math, style recoloring, snapshot consts) to MapSnapshot.kt.
- DashboardDock.kt (590 -> 372 lines): the status cluster and its icon
  helpers to DockStatusCluster.kt.

Cross-file composables flip private -> internal; helpers used by a
single new file stay private in it.
