+++ app/src/main/java/io/github/seijikohara/femto/ui/home/components/MapConfig.kt
@@ -1,0 +1,114 @@
+package io.github.seijikohara.femto.ui.home.components
+
+import android.annotation.SuppressLint
+import android.content.ComponentCallbacks2
+import android.content.Context
+import android.content.res.Configuration
+import android.graphics.Bitmap
+import android.graphics.PixelFormat
+import android.location.Location
+import android.util.Log
+import android.view.SurfaceView
+import android.view.View
+import android.view.ViewGroup
+import androidx.compose.animation.core.Animatable
+import androidx.compose.animation.core.tween
+import androidx.compose.foundation.Canvas
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.isSystemInDarkTheme
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.BoxWithConstraints
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.offset
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.DisposableEffect
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableIntStateOf
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberUpdatedState
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.Path
+import androidx.compose.ui.graphics.TransformOrigin
+import androidx.compose.ui.graphics.asImageBitmap
+import androidx.compose.ui.graphics.drawscope.Stroke
+import androidx.compose.ui.graphics.graphicsLayer
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.platform.LocalDensity
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.IntOffset
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.compose.ui.viewinterop.AndroidView
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.LifecycleEventObserver
+import androidx.lifecycle.compose.LocalLifecycleOwner
+import androidx.lifecycle.repeatOnLifecycle
+import com.composables.icons.lucide.Lucide
+import com.composables.icons.lucide.MapPinOff
+import io.github.seijikohara.femto.R
+import io.github.seijikohara.femto.data.display.MapColorScheme
+import io.github.seijikohara.femto.data.display.MapRenderMode
+import io.github.seijikohara.femto.data.display.MapStyleSetting
+import io.github.seijikohara.femto.ui.theme.FemtoDimens
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.isActive
+import kotlinx.coroutines.suspendCancellableCoroutine
+import org.json.JSONObject
+import org.maplibre.android.MapLibre
+import org.maplibre.android.camera.CameraPosition
+import org.maplibre.android.camera.CameraUpdateFactory
+import org.maplibre.android.geometry.LatLng
+import org.maplibre.android.location.LocationComponentActivationOptions
+import org.maplibre.android.location.modes.CameraMode
+import org.maplibre.android.location.modes.RenderMode
+import org.maplibre.android.maps.MapLibreMap
+import org.maplibre.android.maps.MapLibreMapOptions
+import org.maplibre.android.maps.MapView
+import org.maplibre.android.maps.Style
+import org.maplibre.android.snapshotter.MapSnapshotter
+import kotlin.coroutines.resume
+import kotlin.math.asin
+import kotlin.math.atan2
+import kotlin.math.cos
+import kotlin.math.pow
+import kotlin.math.roundToInt
+import kotlin.math.sin
+
+// User-tunable map rendering config (derived from DisplaySettings): light/dark
+// style, oblique tilt, zoom, the render resolution percent (lower renders a
+// smaller bitmap, faster, upscaled to fill), the user-picked render backend, and
+// the LIVE-only feature toggles (3D buildings / terrain relief).
+internal data class MapConfig(
+    val style: MapStyleSetting = MapStyleSetting.AUTO,
+    val schemeLight: MapColorScheme = MapColorScheme.ACCENT,
+    val schemeDark: MapColorScheme = MapColorScheme.ACCENT,
+    val tiltDeg: Int = 55,
+    val zoom: Int = 16,
+    val renderPercent: Int = 100,
+    val renderMode: MapRenderMode = MapRenderMode.SNAPSHOT,
+    val markerPos: Int = 70,
+    val buildings3d: Boolean = false,
+    val terrain: Boolean = false,
+)
+
+// internal so MapSnapshotRenderTest renders the SAME zoom the panel uses, keeping
+// it a single source of truth (POSITRON_STYLE_URL lives in MapScheme.kt).
+internal const val MAP_ZOOM = 16.5
