+++ app/src/main/java/io/github/seijikohara/femto/ui/home/components/MusicCard.kt
@@ -97,8 +97,8 @@ internal fun MusicCard(
     color = MaterialTheme.colorScheme.surfaceContainer,
 ) {
     when (state) {
-        MusicCardState.NeedsPermission -> ConnectState(onConnect = onConnect)
-        MusicCardState.NoActiveSession -> EmptyState()
+        MusicCardState.NeedsPermission -> MusicConnectState(onConnect = onConnect)
+        MusicCardState.NoActiveSession -> MusicEmptyState()
         is MusicCardState.Playing -> PlayingState(state.nowPlaying, onCommand, onLaunchSource)
     }
 }
@@ -157,396 +157,6 @@ private fun PlayingState(
             modifier = Modifier.fillMaxWidth(),
         )
     }
-}
-
-@Composable
-private fun AlbumArt(
-    nowPlaying: NowPlaying,
-    modifier: Modifier = Modifier,
-) = Box(
-    modifier =
-        modifier
-            .clip(RoundedCornerShape(FemtoDimens.ArtCorner))
-            .background(MaterialTheme.colorScheme.surfaceContainerHigh),
-    contentAlignment = Alignment.Center,
-) {
-    // Crossfade keys on the ImageBitmap identity, which is stable across the
-    // per-second position recompositions, so the dissolve fires only on a real
-    // artwork change (track switch / art arriving) — not on every tick.
-    Crossfade(targetState = nowPlaying.albumArt, label = "albumArt") { art ->
-        if (art != null) {
-            Image(
-                bitmap = art,
-                contentDescription = null,
-                contentScale = ContentScale.Crop,
-                modifier = Modifier.fillMaxSize(),
-            )
-        } else {
-            Box(
-                modifier =
-                    Modifier
-                        .fillMaxSize()
-                        .background(
-                            Brush.linearGradient(
-                                listOf(
-                                    MaterialTheme.colorScheme.primary,
-                                    MaterialTheme.colorScheme.tertiary,
-                                ),
-                            ),
-                        ),
-                contentAlignment = Alignment.Center,
-            ) {
-                Icon(
-                    imageVector = Lucide.Music,
-                    contentDescription = null,
-                    tint = MaterialTheme.colorScheme.onPrimary,
-                    modifier = Modifier.size(48.dp),
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun Meta(
-    source: String,
-    sourceIcon: ImageBitmap?,
-    title: String,
-    artist: String?,
-    album: String?,
-    modifier: Modifier = Modifier,
-) {
-    // Title, artist and album styles derive from the M3 type roles once and are
-    // remembered so a track-change recomposition doesn't reallocate them.
-    val typography = MaterialTheme.typography
-    val titleStyle = remember(typography) { typography.cardTitle() }
-    val secondaryStyle = remember(typography) { typography.cardMeta() }
-    Column(
-        modifier = modifier.fillMaxWidth(),
-        horizontalAlignment = Alignment.Start,
-        verticalArrangement = Arrangement.spacedBy(2.dp),
-    ) {
-        Row(
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.spacedBy(6.dp),
-        ) {
-            // The source app's own icon precedes its name; fall back to a generic
-            // music glyph when the icon could not be resolved (rare).
-            if (sourceIcon != null) {
-                Image(
-                    bitmap = sourceIcon,
-                    contentDescription = null,
-                    modifier = Modifier.size(14.dp).clip(RoundedCornerShape(4.dp)),
-                )
-            } else {
-                Icon(
-                    imageVector = Lucide.Music,
-                    contentDescription = null,
-                    tint = MaterialTheme.colorScheme.onSurfaceVariant,
-                    modifier = Modifier.size(12.dp),
-                )
-            }
-            Text(
-                text = source.uppercase(),
-                style = MaterialTheme.typography.sectionLabel(10, 0.16f),
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-                maxLines = 1,
-            )
-        }
-        MetaLine(
-            icon = Lucide.Music,
-            text = title,
-            style = titleStyle,
-            color = MaterialTheme.colorScheme.onSurface,
-        )
-        MetaLine(
-            icon = Lucide.User,
-            // Radio / stream tracks often carry no artist; show an em dash so the
-            // line height stays stable rather than the title jumping down.
-            text = artist?.takeUnless { it.isBlank() } ?: "—",
-            style = secondaryStyle,
-            color = MaterialTheme.colorScheme.onSurfaceVariant,
-        )
-        // Album is absent for many radio / stream sessions; drop the whole line
-        // rather than show a placeholder, so the block stays compact without it.
-        if (!album.isNullOrBlank()) {
-            MetaLine(
-                icon = Lucide.Disc,
-                text = album,
-                style = secondaryStyle,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-            )
-        }
-    }
-}
-
-@Composable
-private fun MetaLine(
-    icon: ImageVector,
-    text: String,
-    style: TextStyle,
-    color: Color,
-    modifier: Modifier = Modifier,
-) = Row(
-    modifier = modifier,
-    verticalAlignment = Alignment.CenterVertically,
-    horizontalArrangement = Arrangement.spacedBy(6.dp),
-) {
-    Icon(
-        imageVector = icon,
-        contentDescription = null,
-        tint = color,
-        modifier = Modifier.size(14.dp),
-    )
-    Text(
-        text = text,
-        style = style,
-        color = color,
-        maxLines = 1,
-        overflow = TextOverflow.Ellipsis,
-    )
-}
-
-@Composable
-private fun Progress(
-    positionMs: Long,
-    durationMs: Long,
-    positionUpdateTimeMs: Long,
-    isPlaying: Boolean,
-    playbackSpeed: Float,
-) {
-    // Interpolate the displayed position from the PlaybackState basis while
-    // playing, so the bar and elapsed label advance smoothly without waiting for
-    // the next session callback. Held at positionMs when paused.
-    val livePosition by
-        produceState(
-            initialValue = positionMs.coerceIn(0L, durationMs.coerceAtLeast(0L)),
-            positionMs,
-            positionUpdateTimeMs,
-            isPlaying,
-            playbackSpeed,
-            durationMs,
-        ) {
-            val maxMs = durationMs.coerceAtLeast(0L)
-            // Hold the reported position when paused, or when the session gives no
-            // valid update-time basis (a 0 basis would interpolate from boot and
-            // snap the bar to the end).
-            if (!isPlaying || positionUpdateTimeMs <= 0L) {
-                value = positionMs.coerceIn(0L, maxMs)
-                return@produceState
-            }
-            while (true) {
-                val elapsed = SystemClock.elapsedRealtime() - positionUpdateTimeMs
-                value = (positionMs + (elapsed * playbackSpeed)).toLong().coerceIn(0L, maxMs)
-                delay(500L)
-            }
-        }
-    val fraction =
-        if (durationMs > 0L) {
-            (livePosition.toFloat() / durationMs).coerceIn(0f, 1f)
-        } else {
-            0f
-        }
-    Row(
-        modifier = Modifier.fillMaxWidth(),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy(10.dp),
-    ) {
-        Text(
-            text = formatMillis(livePosition),
-            style =
-                MaterialTheme.typography.labelMedium.copy(
-                    fontSize = 11.sp,
-                    fontWeight = FontWeight.Medium,
-                    fontFeatureSettings = TabularFigures,
-                ),
-            color = MaterialTheme.colorScheme.onSurfaceVariant,
-            maxLines = 1,
-        )
-        Box(
-            modifier =
-                Modifier
-                    .weight(1f)
-                    .height(4.dp)
-                    .clip(CircleShape)
-                    .background(MaterialTheme.colorScheme.outlineVariant),
-        ) {
-            Box(
-                modifier =
-                    Modifier
-                        .fillMaxWidth(fraction)
-                        .height(4.dp)
-                        .clip(CircleShape)
-                        .background(MaterialTheme.colorScheme.primary),
-            )
-        }
-        Text(
-            text = formatMillis(durationMs),
-            style =
-                MaterialTheme.typography.labelMedium.copy(
-                    fontSize = 11.sp,
-                    fontWeight = FontWeight.Medium,
-                    fontFeatureSettings = TabularFigures,
-                ),
-            color = MaterialTheme.colorScheme.onSurfaceVariant,
-            maxLines = 1,
-        )
-    }
-}
-
-@Composable
-private fun TransportRow(
-    isPlaying: Boolean,
-    onCommand: (MusicCommand) -> Unit,
-    modifier: Modifier = Modifier,
-) = Row(
-    modifier = modifier,
-    horizontalArrangement = Arrangement.spacedBy(24.dp, Alignment.CenterHorizontally),
-    verticalAlignment = Alignment.CenterVertically,
-) {
-    TransportButton(
-        icon = Lucide.SkipBack,
-        description = stringResource(R.string.music_skip_previous),
-        primary = false,
-        onClick = { onCommand(MusicCommand.SkipPrevious) },
-    )
-    TransportButton(
-        icon = if (isPlaying) Lucide.Pause else Lucide.Play,
-        description = stringResource(R.string.music_play_pause),
-        primary = true,
-        onClick = { onCommand(MusicCommand.PlayPause) },
-    )
-    TransportButton(
-        icon = Lucide.SkipForward,
-        description = stringResource(R.string.music_skip_next),
-        primary = false,
-        onClick = { onCommand(MusicCommand.SkipNext) },
-    )
-}
-
-@Composable
-private fun TransportButton(
-    icon: ImageVector,
-    description: String,
-    primary: Boolean,
-    onClick: () -> Unit,
-) {
-    val width = if (primary) FemtoDimens.MusicPlayButton else FemtoDimens.MusicTransportButton
-    val corner = if (primary) 16.dp else 14.dp
-    val container =
-        if (primary) MaterialTheme.colorScheme.primary else Color.Transparent
-    val content =
-        if (primary) MaterialTheme.colorScheme.onPrimary else MaterialTheme.colorScheme.onSurface
-    Box(
-        modifier =
-            Modifier
-                .height(FemtoDimens.MusicTransportButton)
-                .width(width)
-                .clip(RoundedCornerShape(corner))
-                .background(container)
-                .clickable(onClick = onClick)
-                .semantics { contentDescription = description },
-        contentAlignment = Alignment.Center,
-    ) {
-        Icon(
-            imageVector = icon,
-            contentDescription = null,
-            tint = content,
-            modifier = Modifier.size(32.dp),
-        )
-    }
-}
-
-@Composable
-private fun ConnectState(onConnect: () -> Unit) =
-    Surface(
-        onClick = onConnect,
-        modifier = Modifier.fillMaxSize(),
-        color = MaterialTheme.colorScheme.surfaceContainer,
-    ) {
-        Column(
-            modifier = Modifier.fillMaxSize().padding(FemtoDimens.CardPadding),
-            verticalArrangement = Arrangement.Center,
-            horizontalAlignment = Alignment.CenterHorizontally,
-        ) {
-            Icon(
-                imageVector = Lucide.Music,
-                contentDescription = null,
-                tint = MaterialTheme.colorScheme.primary,
-                modifier = Modifier.size(36.dp),
-            )
-            Box(modifier = Modifier.height(12.dp))
-            Text(
-                text = stringResource(R.string.music_connect_cta),
-                style = MaterialTheme.typography.cardCta(),
-                color = MaterialTheme.colorScheme.onSurface,
-                maxLines = 1,
-            )
-            Box(modifier = Modifier.height(4.dp))
-            Text(
-                // Actionable copy the user must read to unlock the card: clear
-                // the head-unit glance floor (CLAUDE.md#automotive-overrides),
-                // unlike the sanctioned 13sp NoActiveSession EmptyState hint.
-                text = stringResource(R.string.music_connect_hint),
-                style = MaterialTheme.typography.cardCtaHint(),
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-                textAlign = TextAlign.Center,
-                maxLines = 2,
-                overflow = TextOverflow.Ellipsis,
-                modifier = Modifier.widthIn(max = 280.dp),
-            )
-        }
-    }
-
-@Composable
-private fun EmptyState() =
-    Column(
-        modifier = Modifier.fillMaxSize().padding(FemtoDimens.CardPadding),
-        horizontalAlignment = Alignment.CenterHorizontally,
-    ) {
-        // Mockup `.music-card.empty` = `grid-template-rows: 1fr auto auto 1fr`
-        // with `align-content: space-between` and `.empty-icon { align-self:
-        // end }`. The two flexible tracks seat the icon/title/description
-        // cluster at/slightly above the vertical centre, and the icon hugs the
-        // title (its row ends flush against the title row). The taller top
-        // weight nudges the cluster just above centre.
-        Box(modifier = Modifier.weight(1.1f))
-        Icon(
-            imageVector = Lucide.Music,
-            contentDescription = null,
-            tint = MaterialTheme.colorScheme.onSurfaceVariant.copy(alpha = 0.6f),
-            modifier = Modifier.size(36.dp),
-        )
-        Box(modifier = Modifier.height(4.dp))
-        Text(
-            text = stringResource(R.string.music_nothing_playing),
-            style = MaterialTheme.typography.cardCta(),
-            color = MaterialTheme.colorScheme.onSurface,
-            maxLines = 1,
-        )
-        Box(modifier = Modifier.height(4.dp))
-        Text(
-            text = stringResource(R.string.music_nothing_hint),
-            style =
-                MaterialTheme.typography.bodyMedium.copy(
-                    fontSize = FemtoDimens.GlanceTextSize,
-                    lineHeight = 18.sp,
-                ),
-            color = MaterialTheme.colorScheme.onSurfaceVariant,
-            textAlign = TextAlign.Center,
-            maxLines = 2,
-            overflow = TextOverflow.Ellipsis,
-            modifier = Modifier.widthIn(max = 280.dp),
-        )
-        Box(modifier = Modifier.weight(0.9f))
-    }
-
-private fun formatMillis(ms: Long): String {
-    if (ms <= 0L) return "0:00"
-    val totalSeconds = ms / 1000
-    val minutes = totalSeconds / 60
-    val seconds = totalSeconds % 60
-    return "%d:%02d".format(minutes, seconds)
 }
 
 @PreviewLightDark
